Dec20 14h25

+++ run.sh
@@ -1,3 +1,3 @@
-rm *.class
+[ $(ls *class | wc -w) -gt 0 ] && rm *.class
 javac Iliad.java
 java Iliad
